% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_evaluation.R
\name{block_cv_maxent}
\alias{block_cv_maxent}
\title{Cross-validation with geographically masked blocks of occurrence records for evaluation}
\usage{
block_cv_maxent(loc_dat, env_dat, k = 3, coordHeaders = NULL,
  bg_masks = NULL, outputdir, species_name = NULL,
  maxent_settings = list(), eval_metrics = c("auc", "omission_rate",
  "tss"), varying_parameter_name = NULL, do.mask = TRUE,
  do.parallel = TRUE, ncores = NULL, ...)
}
\arguments{
\item{loc_dat}{A two-column matrix,a data.frame, a data.table or a csv file with longitude and latitude coordinates of occurrence records.}

\item{env_dat}{A RasterStack object or a character string specifying the path to a directory with environmental raster layers.}

\item{k}{A numeric integer specifying the number of cluster required.}

\item{coordHeaders}{A character string vector of length two giving the names of the coordinates in \code{loc_dat}}

\item{bg_masks}{A RasterStack object of the background to be partitionned.}

\item{species_name}{[Optional] A character string specifying the name of species. If \code{NULL}, will be set to `Unknown`.}

\item{maxent_settings}{A list of maxent flags.}

\item{do.mask}{A logical}

\item{...}{Additional parameters to be passed to \code{maxent} or \code{make_geographic_block}}
}
\description{
Partition both occurrence records and background into evaluation bins based on some spatial rules.
}
