% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_evaluation.R
\name{calc_omission_rate}
\alias{calc_omission_rate}
\title{utility function that calculates the omission rate from a model for different threshold values}
\usage{
calc_omission_rate(model, occ_train, occ_test, threshold = 10)
}
\description{
utility function that calculates the omission rate from a model for different threshold values
}
