% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_sort.R
\name{chunk2}
\alias{chunk2}
\title{helper function that split a vector x in n chunks of approximatively equal size}
\usage{
chunk2(x, n)
}
\description{
helper function that split a vector x in n chunks of approximatively equal size
}
