% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanOcc.R
\name{cleanOcc}
\alias{cleanOcc}
\title{Clean Species Occurrences}
\usage{

  cleanOcc(species_occ = system.file("extdata/UsefulPlants_workflow/Occ_dir/data_output",
  package = "UsefulPlants"), cutoff_year = 1945,
  cutoff_coord_uncertainty = 20000, buffer_cent = 1000,
  buffer_cap = 10000, buffer_inst = 1000, buffer_gbif_HQ = 1000,
  output_dir = system.file("extdata/Occ_dir/data_output", package =
  "UsefulPlants"), use_TDWG = TRUE, species_id = NULL,
  mc_cores = NULL)
}
\arguments{
\item{species_occ}{Either a directory or a csv file providing occurrence records to clean.}

\item{cutoff_year}{A numeric specifying the cutoff year beyond which records should be removed.}

\item{cutoff_coord_uncertainty}{A numeric specifying the cutoff degree of coordinates uncertainty (in m) beyond which records should be removed.}

\item{buffer_cap}{A numeric specifying the size of the buffer around each capital coordinate (the centre of the city), where records should be ﬂagged as problematic. Default = 10 kilometres}

\item{buffer_inst}{A numeric specifying the size of the buffer around each institution, where records should be ﬂagged as problematic. Default = 1 kilometre}

\item{buffer_gbif_HQ}{A numeric specifying the size of the buffer around the gbif headquarters, where records should be ﬂagged as problematic. Default = 1 kilometre}

\item{output_dir}{A directory where to save species occurrence records data after cleaning.}

\item{use_TDWG}{A logical. Should the occurrences be cleaned by the TDWG regions. Default is \code{TRUE}.}

\item{species_id}{[OPTIONAL] A character string specifying the id of the species in ipni or Kew world checklist database. Only used if \code{use_TDWG=TRUE}.}

\item{buffer_cen}{A numeric specifying the size of the buffer around each province or country centroids, where records should be ﬂagged as problematic. Default = 1 kilometre}

\item{mc_core}{A numeric integer specifying the number of cores to be used for parallel computing. Should be >1 for parallel processing.}
}
\value{
None
}
\description{
Cleans species occurrences using CoordinateCleaner and TDWG region
}
