% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryGBIF.R
\name{get_gbiftaxonkey}
\alias{get_gbiftaxonkey}
\title{Extract Taxon key from GBIF}
\usage{
get_gbiftaxonkey(species_name, rank = "species", kingdom = "Plantae",
  save.outputs = TRUE, index = NULL, check.output = "")
}
\arguments{
\item{species_name}{A vector of species names}

\item{rank}{the given rank at which to search. Default is 'species'.}

\item{kingdom}{the kingdom in which to search. Default is 'Plantae'.}
}
\value{
A numeric value corresponding to the taxon key in the GBIF database
}
\description{
Retrieves the taxon key value of a species from GBIF database
Modified from https://data-blog.gbif.org/post/downloading-long-species-lists-on-gbif/
}
