% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_utilities.R
\name{get_maxent_output}
\alias{get_maxent_output}
\title{From an output directory}
\usage{
get_maxent_output(maxent_outputdir, index = 1)
}
\arguments{
\item{maxent_outputdir}{list object whose elements refer to Maxent arguments}
}
\description{
Reads a named list and returns an executing command line for MaxEnt
}
