% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_richness.R
\name{get_species_richness}
\alias{get_species_richness}
\title{Compute the species richness of geographic units}
\usage{
get_species_richness(species_name = NULL, species_id = NULL,
  status = "native", level = 2, use_name_matching = FALSE,
  do.parallel = TRUE, ncores = NULL, verbose = TRUE)
}
\arguments{
\item{species_name}{A character vector of species names}

\item{species_id}{A character vector of ids either from IPNI or the World Check List [Recommended if available].}

\item{status}{Optional. A character string specifying the type of geographic region where to search for the species. Can be : 'native', 'introduced' or 'both'. Default is 'native'.}

\item{level}{A numeric integer between 1 and 4. The unit level of the TWDG at which the species richness must be calculated. Default is level 2.}

\item{use_name_matching}{A logical. Should the species names be matched with the Kew checklist (i.e. try to guess actual accepted name) ? Default is FALSE.}

\item{do.parallel}{A logical. Should the computation be parallelized ? Default is TRUE}

\item{ncores}{A numeric integer specifying the number of cores to use for parallel computation. Ignored if \code{do.parallel} is set to FALSE.}

\item{verbose}{A logical. Should the name of the species processed be printed on the screen ? Default is TRUE.}
}
\value{
A SpatialPolygonsDataFrame with species richness values stored in the data.
}
\description{
Calculates the species richness at a geographic level
}
