% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{init_main_directories}
\alias{init_main_directories}
\title{Initialize SDM output directories}
\usage{
init_main_directories(base_directory, species_directory, env_directory,
  verbose = TRUE, force_sorting = FALSE, ...)
}
\arguments{
\item{base_directory}{A character string specifying the directory where to store the outputs of the workflow}

\item{species_directory}{A character string specifying the directory where are stored the species occurrence records.}

\item{env_directory}{A character string specifying the directory where are stored the environmental layers.}

\item{verbose}{A logical. Default is `TRUE`.}
}
\value{
None
}
\description{
Creates directories that will contain the outputs of the workflow
}
