% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{init_model_directory}
\alias{init_model_directory}
\title{Initialize model directory}
\usage{
init_model_directory(directory, type = "point", algorithm = "point",
  verbose = FALSE)
}
\arguments{
\item{directory}{A character string specifying the directory where to store the outputs of the model}

\item{type}{A character string specifying the type of model directory to create}

\item{algorithm}{A character string specifying the name of the algorithm used as model}
}
\description{
Creates sub-directories that will contain the outputs of the type of model specified
}
