% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_partitioning.R
\name{kmeansEqual}
\alias{kmeansEqual}
\title{Kmeans clustering with near equal group size}
\usage{
kmeansEqual(kdat, k = 3, iter_max = 30, random.seed = 1234,
  method = c("iqr.outliers", "top.outliers"), nstart = 20,
  n.outliers = 10, plot = TRUE)
}
\arguments{
\item{kdat}{A two-column matrix or data.frame or data.table with longitude and latitude coordinates.}

\item{k}{A numeric integer specifying the number of cluster required.}

\item{iter_max}{A numeric integer specifying the maximum number of iterations to run.}
}
\value{
A list with two elements:
        \code{Data} A two-column matrix or data.frame or data.table with longitude and latitude coordinates and a column 'assigned' specifying
        which of the k cluster each point belongs to.
        \code{Centers} A two-column data.frame with longitude and latitude coordinates of the cluster centers.
}
\description{
Modified from https://rviews.rstudio.com/2019/06/13/equal-size-kmeans/
}
