% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_evaluation.R
\name{loocv_geo}
\alias{loocv_geo}
\title{utility function that performs a leave one out cross validation for geographic models}
\usage{
loocv_geo(pr, bg, tr = seq(from = 0.001, to = 0.99, by = 0.01))
}
\description{
utility function that performs a leave one out cross validation for geographic models
}
