% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_geographic_background.R
\name{make_bbox_domain}
\alias{make_bbox_domain}
\title{Extent the bounding box of a sf point object
https://www.jla-data.net/eng/adjusting-bounding-box-of-a-tmap-map/}
\usage{
make_bbox_domain(loc_dat, coordHeaders = NULL, bbox_ext = 0.5,
  min_xrange = 0.1666667, min_yrange = 0.1666667, left = TRUE,
  right = TRUE, bottom = TRUE, top = TRUE)
}
\arguments{
\item{loc_dat}{A sf object or a two-column matrix or data.frame or data.table with longitude and latitude coordinates.}

\item{coordHeaders}{A character string vector of length two giving the names of the coordinates in \code{loc_dat}}

\item{bbox_ext}{A numeric value specifying how much the bounding box should be extended in percentage.}
}
\description{
Extent the bounding box of a sf point object
https://www.jla-data.net/eng/adjusting-bounding-box-of-a-tmap-map/
}
