% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_partitioning.R
\name{make_geographic_block}
\alias{make_geographic_block}
\title{Geographic "block" partitioning of occurrence records for evaluation}
\usage{
make_geographic_block(kdat, k = 3, bg = NULL,
  method = c("nearest.center"), grid_res = 0.16666667,
  coastline = NULL, ...)
}
\arguments{
\item{kdat}{A two-column matrix or data.frame or data.table with longitude and latitude coordinates.}

\item{k}{A numeric integer specifying the number of cluster required.}

\item{bg}{A RasterLayer object of the background to be partitionned.}

\item{method}{A character string specifying the method to use to assign background cell to each block among "nearest.cluster" and "nearest.center" (the default).}

\item{grid_res}{A numeric value specifying the resolution of the output raster layer in decimal degree. Default is 0.16666667 (~ 20 km).}

\item{coastline}{An optional character string specifying the path to a shapefile delineating the coastline.}

\item{..., }{Additional paramters to be passed to \code{kmeansEqual}}

\item{sf}{A logical. Should a sf object be returned ? Default is TRUE}
}
\value{
A sf object (default) or a RasterStack of the geographically masked blocks of occurrence and background points/cells.
}
\description{
Partition both occurrence records and background into evaluation bins based on some spatial rules.
}
