% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prediction.R
\name{maxent_post_process}
\alias{maxent_post_process}
\title{Post-processing of maxent raw outputs}
\usage{
maxent_post_process(training_raw_maxent, testing_raw_maxent, estim_prev,
  outType = 1, num_bg, H, maxIter = 10)
}
\arguments{
\item{training_raw_maxent}{A vector of maxent raw predictions from training data}

\item{testing_raw_maxent}{A vector of maxent raw predictions from testing data}

\item{estim_prev}{A numeric value estimating the prevalence of the species}

\item{outType}{A integer number between 0-3 specifying which output is to be returned. Default is 0 ('raw', 'logistic' and 'cloglog' not adjusted).
A number between 1-3 involves post-processing: 1 = raw adjusted, 2 = cloglog adjusted, 3 = logistic adjusted.}

\item{num_bg}{A numeric integer specifying the number of background points used for training}

\item{H}{A numeric value specifying the maximum entropy value reached by the model.Only required when \code{outType=1}.}

\item{maxIter}{A numeric integer specifying the number of iterations to run for searching parameter `c` or `tau`. Only require when \code{outType=2} or \code{outType=3}.}
}
\description{
Adjust maxent raw outputs to take into account species prevalence
}
