% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapply.hack.R
\name{mclapply.hack}
\alias{mclapply.hack}
\title{This function mimics forking (done with mclapply in Mac or Linux) for the Windows environment.
Designed to be used just like mclapply.
Code slightly modified from https://rdrr.io/github/lcomm/ltools/ itself (extremely) lightly modified from https://github.com/nathanvan/mcmc-in-irt/blob/master/post-10-mclapply-hack.R
Credit goes to Nathan VanHoudnos.}
\usage{
mclapply.hack(..., verbose = FALSE, mc.cores = NULL)
}
\arguments{
\item{verbose}{Should users be warned this is hack-y? Defaults to FALSE.}
}
\description{
## post-10-mclapply.hack.R
##
## Nathan VanHoudnos
## nathanvan AT northwestern FULL STOP edu
## July 14, 2014
## Last Edit:  August 26, 2014
##
## A script to implement a hackish version of
## parallel:mclapply() on Windows machines.
}
\seealso{
mclapply
}
