% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prediction.R
\name{project_geoModel}
\alias{project_geoModel}
\title{Project a fitted geographic distance-based model}
\usage{
project_geoModel(domain, model, output_directory, output_name = NULL,
  save_output = TRUE)
}
\arguments{
\item{domain}{An object of class `RasterLayer` representing the spatial grid domain where to interpolate the model}

\item{model}{An object of class `GeographicDistance` or `InvDistWeightModel`}

\item{output_directory}{An optional character string specifying the directory where to save the raster of interpolated values}

\item{save_output}{A logical. Should the output raster be saved ? Ignored if `output_directory` is not provided. Default is `TRUE`.}

\item{return}{An object of class `RasterLayer` with values interpolated by the model}
}
\description{
This function is a wrapper for the function `predict` from the `raster` package
}
