% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prediction.R
\name{project_maxent}
\alias{project_maxent}
\title{Project a fitted Maxent model}
\usage{
project_maxent(lambdas, newdata, training_data, outType = 0,
  estim_prevalence, maxIter = 10, return_lfx = FALSE, mask,
  quiet = FALSE)
}
\arguments{
\item{lambdas}{Either (1) a `MaxEnt` fitted model object (fitted with
the `maxent` function in the `dismo` package), (2) a file path to
a Maxent .lambdas file, or (3) a `lambdas` object returned by
[parse_lambdas()].}

\item{newdata}{A `RasterStack`, `RasterBrick`, `list`,
`data.frame`, `data.table`, or `matrix` that has
layers/elements/columns whose names correspond to the names of predictors
used to fit the model. These layers/elements/columns must all have the same
length.}

\item{training_data}{[Optional] Same as newdata. Only required for post-processing options 2 and 3}

\item{outType}{A integer number between 0-3 specifying which output is to be calculated and returned. Default is 0 ('raw', 'logistic' and 'cloglog' not adjusted).
A number between 1-3 involved post-processing: 1 = raw adjusted, 2 = cloglog adjusted, 3 = logistic adjusted.}

\item{estim_prevalence}{A numeric value estimating the prevalence of the species.}

\item{return_lfx}{Logical. Should `Raster` layers be returned giving
lambda*feature values for each feature with a non-zero lambda? Currently
ignored if `newdata` is not a `Raster*` object.}

\item{mask}{(Optional; requires that `newdata` is a `Raster*`
object.) A `Raster` object with `NA` values in cells for which
the model should _not_ be projected. These cells will be assigned
`NA` in the returned output.}

\item{quiet}{Logical. Should projection progress be reported?}
}
\value{
If `newdata` is a `RasterStack` or `RasterBrick`, a
  list with three elements:
* `prediction_raw`: a `Raster` layer giving the raw Maxent
  prediction;
* `prediction_logistic`: a `Raster` layer giving the
  logistic Maxent prediction; and
* `prediction_cloglog`: a `Raster` layer giving the
  cloglog Maxent prediction.
If `newdata` is _not_ a `RasterStack` or `RasterBrick`,
the raster layers will be replaced with `data.table`s in the returned
list.
}
\description{
This function is a wrapper for the function `project` from the `rmaxent` package
It projects a fitted Maxent model to new environmental data enabling for post-processing of maxent predictions.
}
