% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_utilities.R
\name{psi.estim.logistic}
\alias{psi.estim.logistic}
\title{function that implements the logistic output, based on linear predictor 'eta[i]', the relative entropy 'r' and parameter 'tau'}
\usage{
psi.estim.logistic(tau, psi.logi)
}
\description{
function that implements the logistic output, based on linear predictor 'eta[i]', the relative entropy 'r' and parameter 'tau'
}
