% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_utilities.R
\name{psi.estim}
\alias{psi.estim}
\title{function that calculates the mean of the cloglog output, based on linear predictor 'eta' and constant 'c'}
\usage{
psi.estim(c, eta)
}
\description{
function that calculates the mean of the cloglog output, based on linear predictor 'eta' and constant 'c'
}
