% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBIEN.R
\name{queryBIEN}
\alias{queryBIEN}
\title{Query Species Occurrences from BIEN}
\usage{
queryBIEN(species_name, verbose = TRUE)
}
\arguments{
\item{species_name}{A single species name}
}
\value{
A data.table object with occurrence data
}
\description{
Queries species occurrences from BIEN
}
