% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryGBIF.R
\name{queryGBIF}
\alias{queryGBIF}
\title{Query Species Occurrences from GBIF}
\usage{
queryGBIF(species_name, gbif_login, gbif_download_dir = NULL,
  rank = "species", kingdom = "plants", status_ping = 5,
  verbose = TRUE)
}
\arguments{
\item{species_name}{A single species name}

\item{gbif_login}{An object of class GBIFLogin to log in to GBIF to begin the download}
}
\value{
A data.table object with occurrence data
}
\description{
Queries species occurrences from GBIF
}
