% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryOcc.R
\name{queryOcc}
\alias{queryOcc}
\title{Query Species Occurrences}
\usage{

  queryOcc(species_name = system.file("extdata/UsefulPlants_workflow/List_of_useful_plant_name.txt",
  package = "UsefulPlants"), data_sources = c("gbif", "bien", "biotime",
  "rainbio", "genesys", "spLink", "cwr_gbif"),
  download_dir = system.file("extdata/UsefulPlants_workflow/Occ_dir/data_online",
  package = "UsefulPlants"),
  output_dir = system.file("extdata/UsefulPlants_workflow/Occ_dir/data_output",
  package = "UsefulPlants"), run_name = "Test", mc_cores = NULL)
}
\arguments{
\item{species_name}{Either a vector of species names or a csv file where species names are stored in the first column.}

\item{download_dir}{A directory where to save species occurrence records data downloaded from online databases.}

\item{output_dir}{A directory where to save species occurrence records data after formatting.}

\item{mc_core}{A numeric integer specifying the number of cores to be used for parallel computing. Should be >1 for parallel processing.}
}
\value{
None
}
\description{
Queries species occurrences from different databases online and offline
}
