% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{run_geoModel}
\alias{run_geoModel}
\title{Run geographic distance-based model workflow}
\usage{
run_geoModel(loc_dat, species_name = NULL, algorithm = "idw",
  outputdir, output_res = 0.1666667, verbose = TRUE, ...)
}
\arguments{
\item{loc_dat}{A character string specifying the path to a csv file with longitude and latitude coordinates of occurrence records.}

\item{algorithm}{A character string specifying the name of the algorithm to use.}

\item{outputdir}{A character string specifying the directory where to save the raster of interpolated values}

\item{output_res}{A numeric value specifying the resolution of the output raster in decimal degree. Default is 0.1666667 i.e ~ 20km}
}
\value{
An integer number. `0` in case of success otherwise return `-1`
}
\description{
Uses only known occurrence locations to predict the likelihood of finding species in an area surrounding its known presence.
The likelihood of finding the species is thus inversely proportional to the distance to the nearest known presence record.
}
