% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{run_maxentModel}
\alias{run_maxentModel}
\title{Run geographic distance-based model workflow}
\usage{
run_maxentModel(loc_dat, outputdir, verbose = TRUE,
  maxent_settings = list(), ...)
}
\arguments{
\item{loc_dat}{A character string specifying the path to a csv file with longitude and latitude coordinates of occurrence records.}

\item{outputdir}{A character string specifying the directory where to save the raster of interpolated values}

\item{algorithm}{A character string specifying the name of the algorithm to use.}
}
\value{
An integer number. `0` in case of success otherwise return `-1`
}
\description{
Partition both occurrence records and background into evaluation bins based on some spatial rules.
}
