% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{run_pointModel}
\alias{run_pointModel}
\title{Run geographic point-based model}
\usage{
run_pointModel(loc_dat, species_name = NULL, outputdir,
  output_res = 0.1666667, verbose = TRUE, ...)
}
\arguments{
\item{loc_dat}{A character string specifying the path to a csv file with longitude and latitude coordinates of occurrence records.}

\item{outputdir}{A character string specifying the directory where to save the raster of interpolated values}

\item{output_res}{A numeric value specifying the resolution of the output raster in decimal degree. default is 0.1666667 i.e ~ 20km}

\item{algorithm}{A character string specifying the name of the algorithm to use.}
}
\value{
An integer number. `0` in case of success otherwise return `-1`
}
\description{
Uses known occurrence locations as its own.
}
