% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SDMWorkflow.R
\name{run_SDMWorkflow}
\alias{run_SDMWorkflow}
\title{Running SDM Workflow}
\usage{
run_SDMWorkflow(use_category = "All", use_default_settings = TRUE)
}
\arguments{
\item{use_category}{A character string specifying which category of plant uses the modelling should be performed on}

\item{use_default_settings}{A logical. Should the models be run with default settings. Default is TRUE.}
}
\value{
None
}
\description{
Runs Useful Plants Modelling workflow
}
\examples{
run_SDMWorkflow()
}
