% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_utilities.R
\name{search.c}
\alias{search.c}
\title{function to search for the value of 'c' in cloglog that approximates the prevalence}
\usage{
search.c(cL, cH, prev, niter, eta)
}
\description{
function to search for the value of 'c' in cloglog that approximates the prevalence
}
