% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_utilities.R
\name{search.tau}
\alias{search.tau}
\title{function to search for the value of 'tau' that approximates the observed prevalence given in prev}
\usage{
search.tau(tauL, tauH, prev, niter, psi.logi)
}
\description{
function to search for the value of 'tau' that approximates the observed prevalence given in prev
}
