% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_sort.R
\name{sortByInitNumRecords}
\alias{sortByInitNumRecords}
\title{Sort species by number of occurrence records retrieve from occurrence data files.}
\usage{
sortByInitNumRecords(base_directory, species_directory, env_directory,
  algorithm, min_occ_envmodel = 10, min_occ_geomodel = 3,
  nchunk = 20, ncores = NULL, verbose = FALSE)
}
\arguments{
\item{base_directory}{A character string specifying the directory where to store the outputs of the workflow}

\item{species_directory}{A character string specifying the directory where are stored the species occurrence records.}

\item{env_directory}{A character string specifying the directory where are stored the environmental layers.}

\item{min_occ_envmodel}{A numeric integer specifying the minimum number of occurrence records required to assign a species to the environmental model directory}

\item{min_occ_geomodel}{A numeric integer specifying the minimum number of occurrence records required to assign a species to the geographic model directory}

\item{nchunk}{A numeric integer specifying the number of chunks in which the vector of species names must be divided. Must be > 1 to enable parallel processing.}

\item{ncores}{A numeric integer specifying the number of cores (or CPU's) to use in case of parallel processing. Default is `number of cores available` - 1.}
}
\description{
Assign species occurrence data to different folders based on minimum number of occurrence records
}
