% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_sorting_bias.R
\name{split2}
\alias{split2}
\title{helper function that split a vector x in chunks of maximum size d}
\usage{
split2(x, d)
}
\description{
helper function that split a vector x in chunks of maximum size d
}
