% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsefulPlants.R
\name{usefulplants}
\alias{usefulplants}
\title{Utilised plant species data}
\format{
A data.table containing a list of 35687 plant species names and 10 plant use categories.

\describe{
  \item{binomial_acc_name}{binomial accepted name of the plant species}
  \item{author_acc_name}{author accepted name(s)}
  \item{AnimalFood}{EnvironmentalUses}{Fuels}{GeneSources}{HumanFood}{InvertebrateFood}{Materials}{Medicines}{Poisons}{SocialUses}{name of plant uses}
  \item{Totals}{total number of uses reported for a species}

}
}
\source{
\url{https://zenodo.org/record/8176318}
}
\description{
List of plant species used by humans
}
