% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_utilities.R
\name{maXent}
\alias{maXent}
\title{Performing Species Distribution Modelling with MaxEnt}
\usage{
maXent(sp_points, env_layers, outputdir, path_to_maxent = NULL,
  path_to_java = NULL, wait = TRUE, ...)
}
\arguments{
\item{sp_points}{A character string specifying the path to the species occurrence csv files}

\item{env_layers}{A character string specifying the path to the directory containing environmental predictors layers}

\item{outputdir}{A character string specifying the path to the output directory for maxent results.}

\item{path_to_maxent}{A character string specifying the path to the maxent .jar file}

\item{path_to_java}{A character string specifying the path to a java interpreter}

\item{wait}{A logical. Should the function wait until the job is finished ? Default is TRUE.}

\item{...}{Additional parameters to be passed to maxent.}
}
\description{
Runs MaxEnt on a species occurrence dataset and with given environmental variables
}
