% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryGBIF.R
\name{queryGBIF}
\alias{queryGBIF}
\title{Query Species Occurrences from GBIF}
\usage{
queryGBIF(species_name, user = NULL, email = NULL, pwd = NULL,
  gbif_download_dir = NULL, rank = "species", kingdom = "Plantae",
  status_ping = 3, time_out = 300, index = NULL, check_output = "",
  verbose = TRUE)
}
\arguments{
\item{species_name}{A vector of species names}

\item{user}{A vector of type character specifying a GBIF username.}

\item{email}{A vector of type character specifying the email associated with a GBIF username.}

\item{pwd}{A vector of type character containing the user's password for logging in to GBIF.}

\item{rank}{the given rank at which to search. Default is 'species'.}

\item{kingdom}{the kingdom in which to search. Default is 'Plantae'.}
}
\value{
A data.table object with occurrence data
}
\description{
Queries species occurrences from GBIF
}
