% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SDMWorkflow.R
\name{run_SDMWorkflow}
\alias{run_SDMWorkflow}
\title{Running SDM Workflow}
\usage{
run_SDMWorkflow(base_directory, species_directory, env_directory,
  which_uses = "AllUses", use_default_settings = TRUE)
}
\arguments{
\item{base_directory}{A character string specifying the directory where to store the outputs of the workflow}

\item{species_directory}{A character string specifying the directory where are stored the species occurrence records.}

\item{env_directory}{A character string specifying the directory where are stored the environmental layers.}

\item{which_uses}{A character string specifying the name of the plant use category to model among: `AnimalFood`, `HumanFood`,`InvertebratesFood`,`Medicinals`
`EnvironmentalUses`,`Materials`,`SocialUses`,`GeneSources`,`Fuels`,`Poisons`. Default is `AllUses`, which means that outputs of all categories will be aggregated.}
}
\value{
None
}
\description{
Runs Useful Plants Modelling workflow
}
\examples{
run_SDMWorkflow()
}
